Split marm-console's server/app.py into models + core + endpoints/
Pure refactor, no behavior changes, no route/path/method changes.
Mirrors marm-mcp-server's own core/models.py + endpoints/*.py + router
registration convention (confirmed with user over a lighter grouping).
Extracted via exact index-based line-slicing -- all 15 pieces (including
4 non-contiguous ones) verified byte-identical to the original before
any import cleanup.

- server/app.py keeps only app wiring: CORS, startup event, /health,
  and 8 app.include_router(...) calls. 915 -> 74 lines.
- server/models.py (new): all 15 Pydantic payload models.
- server/core.py (new): helpers genuinely shared across 2+ domains --
  _now_iso, get_memory_db_path, get_concept_db_path, _concepts_payload,
  _mcp_tool_mutation.
- server/endpoints/ (new package, 8 files): overview, memory, sessions,
  logs, notebook, compaction, concepts, projects. Each owns its
  domain's routes plus any helper used only within that domain
  (_memory_mutation stays in memory.py, _project_operation stays in
  projects.py, concept-build job-tracking state stays in concepts.py).

Deliberate departure from a literal 1:1 mirror: /health, CORS, and the
startup event stay in app.py rather than getting their own
endpoints/system.py -- one 3-line endpoint with zero shared deps would
be exactly the "tiny module with no real boundary" refactor-middle-
ground.md warns against.

Two real bugs caught during implementation, both by tooling working as
designed:
- ruff check caught endpoints/concepts.py calling _concepts_payload
  without importing it from core.py.
- pytest caught a gap ruff couldn't: test_memory_dashboard_gap_routes.py
  monkeypatched console_app.memory_store directly (4 call sites), which
  broke once app.py correctly stopped importing memory_store itself.
  Repointed to import memory_store directly from server -- also fixed
  test_concept_store.py's console_app._stale_build_result call the same
  way, since that function moved to endpoints/concepts.py.

Validated: ruff check + ruff format --check clean, full suite 18
passed / 0 failed (matches baseline), and a live TestClient smoke pass
confirmed all 8 domain routers plus /health are correctly mounted (no
silent 404s from a missing include_router call).
