+++ marm-console/server/endpoints/__init__.py
@@ -1,0 +1,1 @@
+"""Domain-grouped REST endpoint routers for MARM Console."""
